ai_agent/requirements.txt
web3
google-generativeai
python-dotenv
